Added the undo stack to Graphical Recording Mode
This patch introduces the concept of the "undo stack" to graphical
recording mode.

Basically, we support a set list of "actions" that can be undone by
pressing the undo button in the UI.

This replaces the old concept of undoing a move, which simply removed
the last move in the move history list.

This first implementation only supports undoing a "toggle draw offer"
action - subsequent patches will add more support